GH-7 - Restructure into a multi module project.
This also removes the need for including binaries, at least partially. I wasn’t able to configure the test-jar plugin nor the invoker plugin to create an artefact from `test-migrations`that was actual referencable from the `core` module. During direct maven build of the parent project, the dependencie will be provided as jar, so the test for resolvable migrations in jars will fit it’s purpose.

If a known resource is not available under the `jar` protocol, than the test in question is skipped.

This closes #7.