fix: Also catch `Neo.ClientError.Schema.ConstraintAlreadyExists` while locking. (#324)
A `Neo4jException` is thrown with the code `Neo.ClientError.Schema.ConstraintAlreadyExists` when a constraint already exists that is not only equivalent but has a name.
The same scenario might happen in an upgrade from 4.3 to 4.4. at least, as there the mechanism that decides which constraints are equivalent seems to have changed.

Thanks to @ali-ince for reporting this and help me fixing it.