feat: Allow disabling of migrations in Spring Boot without removing the `Migrations` bean. (#361)
This is a breaking change in behavior and aligns the behavior of the Spring Boot starter with the Quarkus extension: Disabling migrations will just disable migrations on startup but still provide a `Migrations` bean with the given configuration. This allows for calls to `info`, `validate` and the like without manually recreating the configuration.

This closes #342.