feat: Support classpath scanning in Quarkus native image.
This turns locations- and packages to scan into build time configuration options and includes discovered classpath resources into the native image, allowing the Quarkus extension to be fully usable in a native image.

To make up for the loss of dynamically adding external resources, a `external-locations` runtime property has been added.

This closes #362.