fix: Improve parsing of single line comments.
The spliting of statements works on `;` followed by a linebreak as indicator. That leads to the parser not recognizing multiple single line comments without a following statement as single line comments.

We cannot change the splitting of statements as that would change the checksum which is computed on a per statement and not per file basis. Therefore the same algorithm like in `getSingleLineComments` has been applied now to filter those.