refactor: Warn only on empty locations. (#555)
Maven and other build tools won't create empty folders in the target unless there are files stores in it. As a result, a project with `neo4j-migrations-spring-boot` will fail on start-up unless there are some actual migrations files are provided. The old decision has been inconvenient and frustrating, so we change this to a warning log.