build: Package the modules prior to site generation.
The manifest plugin adds the automic (but deterministic) module name in the packaging phase, not in the compile phase.
Therefor, the site plugin would fail with

```
Creating an aggregated report for both named and unnamed modules is not possible.
Ensure that every module has a module descriptor or is a jar with a MANIFEST.MF containing an Automatic-Module-Name.
```

as the core module now is on the module path by default.