feat: Allow programmatic control over repeatable Java based migrations.
This allows to mark Java based migrations to be repeatable unrelated to their version name and also makes them repeat only when their checksum changes.

This is a breaking change if you relied on the fact that Java based migrations are always repeated! If you need this, return a random checksum on each call of `getChecksum` or a different one in case you actually want the migration to be repeated.

Close #705.