build: Improve Antora includes and build
Improve inclusion of examples and also makes the current released version appear in the docs (as of writing 1.14.0), instead of an unreleased one and keeping the prerelease info and the Antora version. This is slightly incorrect but at least all the Gradle / Maven and other examples spotting a version number will work.

Also make the notification workflow run.

nocll