refactor: Add full testing support for Neo4j 5.
This is partially from https://github.com/michael-simons/neo4j-migrations/pull/699 which needs rebasing.
Summary: There is a 4.0 ARM version, we need to deal with the faster release pace of Neo4j 5 (it is not sustainable to add a new enum value each month) and we also want to have `LATEST` support in the image name substitutor.