feat: Provide a limited set of annotations to define constraints. (#755)
This change adds a new module called `eu.michael-simons.neo4j:neo4j-migrations-annotation-catalog` with `@Required` and `@Unique` that can be used together with Spring Data Neo4j or standalone to shape your constraints around a class model.

The existing annotation processor will deal with both annotations and write catalogs accordingly.

While there is a broad range of annotations that are somewhat similar to `@Required`, non of them allows without further additions to define labels or types as the target of the constraint to be created. Therefore we decided on a custom one. XKCD 927 applies.

Closes #703