fix: Apidocs generation was not running since 2.0.0.
The obvious cause was using the `fast` profile during build, which caused the JavaDoc plugin to be skipped. The `package` however is required, as JavaDoc does not allow named and unnamed modules on the path. We don’t have unnamed modules, but modules with an automatic name, added via maven-jar plugin, which is only available for javadoc via `javadoc:aggregate`. The latter was already defined in the reporting, having the standard `javadoc` there is wrong and would lead to missing modules exceptions.

The Spring Boot examples have some issues, but I decided to just skip the JavaDocs for them.

The JavaDocs for the Maven plugin must be skipped as there’s one automatic module on the classpath (plexus container default), which has a very invalid module name.