fix: Adapt to Spring Boot 3.2.0 revamped support for nested jars.
This fixes #1153 by checking for `jar:file` plus a nested path when a `FileNotFoundException` is thrown when opening a jar file. If those are present, we assume Boot 3.2 with non-classic loader and reshape the URL into the new format.

For now, Classgraph does not recognize the new format (see https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-3.2-Release-Notes#nested-jar-support), but at least finds those resources.

This adds a nested Jar to the Spring Boot example tests, too.