fix: Check for `!/BOOT-IN/` vs `/!BOOT-INF/`at all relevant places of resource loading.
The `ResourceContext` now has a centralised way of getting an `InputStream` for resources and checks if it runs on a Spring Boot > 3.2.0 modern nested Jar.
The factory methods for the Cypher resources that take a raw URL have been deprecated and will be removed in 3.x.
All other factory methods of resources have been internal anyway.

Fixes #1248.