feat: Add support for custom auth tokens.
This change allows the CLI to accept custom auth tokens.

Custom auth tokens are tried after all password options and after looking for a bearer token.

Custom auth tokens are specified via the `—custom-auth` CLI option. Several entries are given like this:

```
--custom-auth=principal=neo4j
--custom-auth=credentials=verysecret
--custom-auth=scheme=basic
—custom-auth=realm=whatever
—custom-auth=additionalentry=value
```

Every entry in that map not matching `principal`, `credentials`, `scheme`or `realm` will be passed as is to the auth token.

Closes #1272.