feat: Add support for configuring the transaction timeout. (#1430)
This adds  a `transactionTimeout` config option in the core API and all supported extensions. It will be passed on towards the driver, and might be `null`, the latter indicating the drivers default.

The timeout is applied either individually to each migration script or per statement in a script, depending on the `transactionMode`. 

Refer to the drivers manual with regard to the implications of transaction time out https://neo4j.com/docs/java-manual/current/ as well as general transaction management https://neo4j.com/docs/operations-manual/current/database-internals/transaction-management/

This closes #1428.