feat: Support relative files paths.
This adds kind of a dangerous feature: Location parsing has been enhanced so that locations without a `classpath` or `file` prefix, starting with either a single dot or two consecutive dots are treated as relative locations against the local working directory.

This feature has not been added to the `file` protocol, as that will always include the `/`, denotating the root of the filesystem (see https://en.wikipedia.org/wiki/File_URI_scheme), and for every relative component in it, I think it’s safer to normalize them away.

Also, any directory that does not exist but is configured is now logged on warning level.

This closes #1767

Signed-off-by: Michael Simons <michael@simons.ac>